feat(point): 포인트 충전 기능 구현 -로그인한 사용자인지 확인 -세션을 통해 로그인 사용자 데이터 사용

+++ src/main/java/com/geonwoo/solokill/domain/payment/model/Payment.java
@@ -1,6 +1,5 @@
 package com.geonwoo.solokill.domain.payment.model;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.geonwoo.solokill.domain.member.model.Member;
@@ -25,11 +24,11 @@ public class Payment {
 	@ManyToOne
 	private Member member;
 
-	private BigDecimal value;
+	private Long value;
 
 	private LocalDateTime createAt;
 
-	public Payment(Member member, BigDecimal value) {
+	public Payment(Member member, Long value) {
 		this.member = member;
 		this.value = value;
 		this.createAt = LocalDateTime.now();
